+++ src/nowiki/list.ts
@@ -4,7 +4,7 @@ import {syntax} from '../../mixin/syntax';
 import {ListBaseToken} from './listBase';
 
 /** 位于行首的`;:*#` */
-@sol(true) @syntax(/^[;:*#][:*#]*[^\S\n]*$/u)
+@sol(true) @syntax(/^[;:*#]+[^\S\n]*$/u)
 export abstract class ListToken extends ListBaseToken {
 	override get type(): 'list' {
 		return 'list';